add z option handle to playlist-archive-rm

+++ init.sh
@@ -6,12 +6,18 @@ timestamp="$(date +%s)"
 echo ===untrash===
 "$scriptdir/untrash.sh"
 if [ -f ~/Music/maybe\ remove.m3u ]; then
-echo ===fromplaylist\|grep archivedir\|rm===
-grep -v '^#' ~/Music/maybe\ remove.m3u | grep -F "/$(basename "$(cat "$scriptdir/archivedir")")/" | sed 'ss^.*/\([^/]*/[^/]*\)$s\1s' |
-	xargs -rd \\n -n 1 printf '%s/%s\n' "$(cat "$scriptdir/archivedir")" | xargs -rd \\n rm -v --
-echo ===fromplaylist\|toignore===
-grep -v '^#' ~/Music/maybe\ remove.m3u | "$scriptdir/toignore.sh"
-cat ~/Music/maybe\ remove.m3u 2>&1 1>> ~/Music/maybe\ remove~.m3u && rm ~/Music/maybe\ remove.m3u
+	if [ "$1" = "z" ]; then
+		echo ===fromplaylist\|grep archivedir\|rm===
+		grep -v '^#' ~/Music/maybe\ remove.m3u | grep -F "/$(basename "$(cat "$scriptdir/archivedir")")/" | sed 'ss^.*/\([^/]*/[^/]*\)$s\1s' |
+			xargs -rd \\n -n 1 printf '%s/%s\n' "$(cat "$scriptdir/archivedir")" | xargs -rd \\n rm -v --
+	else
+		echo ===fromplaylist\|grep archivedir===
+		grep -v '^#' ~/Music/maybe\ remove.m3u | grep -F "/$(basename "$(cat "$scriptdir/archivedir")")/" | sed 'ss^.*/\([^/]*/[^/]*\)$s\1s' |
+			xargs -rd \\n -n 1 printf '%s/%s\n' "$(cat "$scriptdir/archivedir")"
+	fi
+	echo ===fromplaylist\|toignore===
+	grep -v '^#' ~/Music/maybe\ remove.m3u | "$scriptdir/toignore.sh"
+	cat ~/Music/maybe\ remove.m3u 2>&1 1>> ~/Music/maybe\ remove~.m3u && rm ~/Music/maybe\ remove.m3u
 fi
 
 echo ===dl\&recentinlog===
